Changes in DefaultParams from int to ulong

+++ net/dll/seal/defaultparams.cpp
@@ -11,7 +11,7 @@ using namespace seal;
 using namespace seal::dll;
 
 
-SEALDLL HRESULT SEALCALL DefParams_CoeffModulus128(int polyModulusDegree, int* length, void** coeffs)
+SEALDLL HRESULT SEALCALL DefParams_CoeffModulus128(uint64_t polyModulusDegree, uint64_t* length, void** coeffs)
 {
     IfNullRet(length, E_POINTER);
 
@@ -29,7 +29,7 @@ SEALDLL HRESULT SEALCALL DefParams_CoeffModulus128(int polyModulusDegree, int* l
     return S_OK;
 }
 
-SEALDLL HRESULT SEALCALL DefParams_CoeffModulus192(int polyModulusDegree, int* length, void** coeffs)
+SEALDLL HRESULT SEALCALL DefParams_CoeffModulus192(uint64_t polyModulusDegree, uint64_t* length, void** coeffs)
 {
     IfNullRet(length, E_POINTER);
 
@@ -47,7 +47,7 @@ SEALDLL HRESULT SEALCALL DefParams_CoeffModulus192(int polyModulusDegree, int* l
     return S_OK;
 }
 
-SEALDLL HRESULT SEALCALL DefParams_CoeffModulus256(int polyModulusDegree, int* length, void** coeffs)
+SEALDLL HRESULT SEALCALL DefParams_CoeffModulus256(uint64_t polyModulusDegree, uint64_t* length, void** coeffs)
 {
     IfNullRet(length, E_POINTER);
 
@@ -65,7 +65,7 @@ SEALDLL HRESULT SEALCALL DefParams_CoeffModulus256(int polyModulusDegree, int* l
     return S_OK;
 }
 
-SEALDLL HRESULT SEALCALL DefParams_SmallMods60Bit(int index, void** small_modulus)
+SEALDLL HRESULT SEALCALL DefParams_SmallMods60Bit(uint64_t index, void** small_modulus)
 {
     IfNullRet(small_modulus, E_POINTER);
 
@@ -81,7 +81,7 @@ SEALDLL HRESULT SEALCALL DefParams_SmallMods60Bit(int index, void** small_modulu
     }
 }
 
-SEALDLL HRESULT SEALCALL DefParams_SmallMods50Bit(int index, void** small_modulus)
+SEALDLL HRESULT SEALCALL DefParams_SmallMods50Bit(uint64_t index, void** small_modulus)
 {
     IfNullRet(small_modulus, E_POINTER);
 
@@ -97,7 +97,7 @@ SEALDLL HRESULT SEALCALL DefParams_SmallMods50Bit(int index, void** small_modulu
     }
 }
 
-SEALDLL HRESULT SEALCALL DefParams_SmallMods40Bit(int index, void** small_modulus)
+SEALDLL HRESULT SEALCALL DefParams_SmallMods40Bit(uint64_t index, void** small_modulus)
 {
     IfNullRet(small_modulus, E_POINTER);
 
@@ -113,7 +113,7 @@ SEALDLL HRESULT SEALCALL DefParams_SmallMods40Bit(int index, void** small_modulu
     }
 }
 
-SEALDLL HRESULT SEALCALL DefParams_SmallMods30Bit(int index, void** small_modulus)
+SEALDLL HRESULT SEALCALL DefParams_SmallMods30Bit(uint64_t index, void** small_modulus)
 {
     IfNullRet(small_modulus, E_POINTER);
